Add a start offset for the counter to mkblocks.sh
This feature adds the ability to designate a start offset for the
counter that creates the spf blocks. Currently, the counter starts
at 1, which will always create an entry for `spf.example.com`. If
it is desirable to start the numbering with `spf1.example.com`,
then supplying `mkblocks.sh -s 2` will satisfy this requirement.

This need arose because our `example.com` TXT record cannot be changed
programatically, so specifying `v=spf1 include:spf1.example.com ~all`
in the `example.com` TXT record will allow the `spf#.example.com`
records to be updated independently.

Tests are also included for this new functionality.